created test_db.py file
